requirements.txt
Django==3.2.20
djangorestframework==3.14.0
django-cors-headers==4.3.0
djongo==1.3.6
pymongo==3.12.3
python-dotenv==1.0.0
sqlparse==0.2.4 